Did a lot of things, still have a lot of problems.
Problem with rotations on animation, giving me negative rotation.

Moved bodies in rube so that the fixutre origin is good to rotate with animation.
(fixture is seperate from body, press f in rube to edit fixtures, b for bodies)

Added collision filters for arms and legs so they don't collide.
Category = category
Mask = which category it collides with

Collision filters:
https://www.youtube.com/watch?v=rGJQe0vtIo0