Made using physics more secure against crashes. Instead of using getWorld() to call physics methods such as createBody() or destroyJoint() Physics has it's own implementation of these methods with error checks to prevent crashes which do not give any info on the actual problem.
Fixed a crashing issue where there would be a crash because of the
exessive calls to stopGetUp() and the fact that stopGetUp() would only stop getup
when getup was in progress instead of preventing it.

using camera.unproject(camera.project()) in conjunction like this does not work.

camera.project() doesn't take into account the fact that screenY starts at top.
camera.unproject() does take that into account.

This is why when doing this you have to do camera.viewportHeight - pos.y